Extract Commands from Options module
