+++ lib/fine-types/app/Commands/Check.hs
@@ -1,0 +1,8 @@
+module Commands.Check where
+
+import Prelude
+
+import Options.Check (CheckOptions)
+
+check :: CheckOptions -> IO ()
+check _ = error "Not implemented"
