+++ lib/fine-types/fine-types.cabal
@@ -129,12 +129,17 @@ executable fine-types
     , aeson-pretty
     , base
     , bytestring
+    , contra-tracer
     , megaparsec
     , optparse-applicative
     , fine-types
     , with-utf8
     , yaml
   other-modules:
+    Commands.Check
+    Commands.Common
+    Commands.Convert
+    Commands.Log
     Options
     Options.Check
     Options.Common
